fixed revisions

+++ lib/screens/pages/customer/notif_page.dart
@@ -83,7 +83,8 @@ class CustomerNotifPage extends StatelessWidget {
                           child: ListView.builder(
                             itemCount: data.docs.length,
                             itemBuilder: (context, index) {
-                              double points = data.docs[index]['pts'];
+                              double points =
+                                  data.docs[index]['pts'].toDouble();
                               return Padding(
                                 padding: const EdgeInsets.all(5.0),
                                 child: Card(
